chore: Cleanup app layout unit tests
