+++ src/app-layout/__tests__/drawers.test.tsx
@@ -8,14 +8,7 @@ import { KeyCode } from '@cloudscape-design/test-utils-core/utils.js';
 import AppLayout, { AppLayoutProps } from '../../../lib/components/app-layout';
 import createWrapper from '../../../lib/components/test-utils/dom';
 import { testIf } from '../../__tests__/utils';
-import {
-  describeEachAppLayout,
-  findActiveDrawerLandmark,
-  manyDrawers,
-  manyDrawersWithBadges,
-  renderComponent,
-  testDrawer,
-} from './utils';
+import { describeEachAppLayout, manyDrawers, renderComponent, testDrawer } from './utils';
 
 import toolbarTriggerButtonStyles from '../../../lib/components/app-layout/visual-refresh-toolbar/toolbar/trigger-button/styles.css.js';
 import visualRefreshStyles from '../../../lib/components/app-layout/visual-refresh/styles.selectors.js';
@@ -32,6 +25,55 @@ jest.mock('@cloudscape-design/component-toolkit', () => ({
 }));
 
 describeEachAppLayout(({ size, theme }) => {
+  test('registers public drawers api', () => {
+    const { wrapper } = renderComponent(<AppLayout drawers={[testDrawer]} />);
+    expect(wrapper.findDrawersTriggers()).toHaveLength(1);
+  });
+
+  test('should render an active drawer', () => {
+    const { wrapper } = renderComponent(
+      <AppLayout activeDrawerId={testDrawer.id} drawers={[testDrawer]} onDrawerChange={() => {}} />
+    );
+
+    expect(wrapper.findActiveDrawer()).toBeTruthy();
+  });
+
+  test('should call handler once on open when toggle is clicked', () => {
+    const onChange = jest.fn();
+    const { wrapper } = renderComponent(
+      <AppLayout drawers={[testDrawer]} onDrawerChange={event => onChange(event.detail)} />
+    );
+
+    wrapper.findDrawerTriggerById(testDrawer.id)!.click();
+    expect(onChange).toHaveBeenCalledWith({ activeDrawerId: 'security' });
+  });
+
+  test('should call handler once on open when span inside toggle is clicked', () => {
+    const onChange = jest.fn();
+    const { wrapper } = renderComponent(
+      <AppLayout drawers={[testDrawer]} onDrawerChange={event => onChange(event.detail)} />
+    );
+
+    wrapper.findDrawerTriggerById('security')!.find('span')!.click();
+    expect(onChange).toHaveBeenCalledTimes(1);
+    expect(onChange).toHaveBeenCalledWith({ activeDrawerId: 'security' });
+  });
+
+  test('should call handler once on close', () => {
+    const onChange = jest.fn();
+    const { wrapper } = renderComponent(
+      <AppLayout
+        activeDrawerId={testDrawer.id}
+        drawers={[testDrawer]}
+        onDrawerChange={event => onChange(event.detail)}
+      />
+    );
+
+    wrapper.findActiveDrawerCloseButton()!.click();
+    expect(onChange).toHaveBeenCalledTimes(1);
+    expect(onChange).toHaveBeenCalledWith({ activeDrawerId: null });
+  });
+
   test(`should not render drawer when it is not defined`, () => {
     const { wrapper, rerender } = renderComponent(<AppLayout toolsHide={true} drawers={[testDrawer]} />);
     expect(wrapper.findDrawersTriggers()).toHaveLength(1);
@@ -65,49 +107,6 @@ describeEachAppLayout(({ size, theme }) => {
     expect(wrapper.findActiveDrawer()).toBeTruthy();
   });
 
-  test('renders correct aria-label on overflow menu', () => {
-    const ariaLabels: AppLayoutProps.Labels = {
-      drawersOverflow: 'Overflow drawers',
-      drawersOverflowWithBadge: 'Overflow drawers (Unread notifications)',
-    };
-    const { wrapper, rerender } = renderComponent(<AppLayout drawers={manyDrawers} ariaLabels={ariaLabels} />);
-    const buttonDropdown = wrapper.findDrawersOverflowTrigger();
-
-    expect(buttonDropdown!.findNativeButton().getElement()).toHaveAttribute('aria-label', 'Overflow drawers');
-
-    rerender(<AppLayout drawers={manyDrawersWithBadges} ariaLabels={ariaLabels} />);
-    expect(buttonDropdown!.findNativeButton().getElement()).toHaveAttribute(
-      'aria-label',
-      'Overflow drawers (Unread notifications)'
-    );
-  });
-
-  test('overflow menu item have aria-role set to `menuitem`', () => {
-    const { wrapper } = renderComponent(<AppLayout drawers={manyDrawers} />);
-    const buttonDropdown = wrapper.findDrawersOverflowTrigger();
-
-    buttonDropdown!.openDropdown();
-
-    const countItems = buttonDropdown!.findItems();
-    const countRoleMenuItemRole = buttonDropdown!
-      .findOpenDropdown()!
-      .find('[role="menu"]')!
-      .findAll('[role="menuitem"]');
-
-    expect(countItems.length).toBe(countRoleMenuItemRole.length);
-  });
-
-  test('renders aria-labels', () => {
-    const { wrapper } = renderComponent(<AppLayout drawers={[testDrawer]} />);
-    expect(wrapper.findDrawerTriggerById('security')!.getElement()).toHaveAttribute(
-      'aria-label',
-      'Security trigger button'
-    );
-    wrapper.findDrawerTriggerById('security')!.click();
-    expect(findActiveDrawerLandmark(wrapper)!.getElement()).toHaveAttribute('aria-label', 'Security drawer content');
-    expect(wrapper.findActiveDrawerCloseButton()!.getElement()).toHaveAttribute('aria-label', 'Security close button');
-  });
-
   test('renders resize only on resizable drawer', () => {
     const { wrapper } = renderComponent(
       <AppLayout
@@ -126,15 +125,7 @@ describeEachAppLayout(({ size, theme }) => {
     expect(wrapper.findActiveDrawerResizeHandle()).toBeFalsy();
 
     wrapper.findDrawerTriggerById('security-resizable')!.click();
-    if (size === 'desktop') {
-      expect(wrapper.findActiveDrawerResizeHandle()).toBeTruthy();
-      expect(wrapper.findActiveDrawerResizeHandle()!.getElement()).toHaveAttribute(
-        'aria-label',
-        'Security resize handle'
-      );
-    } else {
-      expect(wrapper.findActiveDrawerResizeHandle()).toBeFalsy();
-    }
+    expect(!!wrapper.findActiveDrawerResizeHandle()).toBe(size === 'desktop');
   });
 
   test('focuses drawer close button', () => {
@@ -178,11 +169,6 @@ describeEachAppLayout(({ size, theme }) => {
     expect(wrapper.findActiveDrawerCloseButton()!.getElement()).toHaveFocus();
   });
 
-  test('registers public drawers api', () => {
-    const { wrapper } = renderComponent(<AppLayout drawers={[testDrawer]} />);
-    expect(wrapper.findDrawersTriggers()).toHaveLength(1);
-  });
-
   testIf(size !== 'mobile')('aria-controls points to an existing drawer id', () => {
     const { wrapper } = renderComponent(<AppLayout drawers={[testDrawer]} />);
     const drawerTrigger = wrapper.findDrawerTriggerById('security')!;
@@ -191,6 +177,31 @@ describeEachAppLayout(({ size, theme }) => {
     drawerTrigger.click();
     expect(drawerTrigger!.getElement()).toHaveAttribute('aria-controls', 'security');
     expect(wrapper.findActiveDrawer()!.getElement()).toHaveAttribute('id', 'security');
+  });
+
+  test('should render badge when defined', () => {
+    const { wrapper } = renderComponent(
+      <AppLayout
+        drawers={[
+          { ...testDrawer, id: 'first', badge: true },
+          { ...testDrawer, id: 'second' },
+        ]}
+      />
+    );
+
+    expect(wrapper.findDrawerTriggerById('first', { hasBadge: true })).toBeTruthy();
+    expect(wrapper.findDrawerTriggerById('second', { hasBadge: false })).toBeTruthy();
+  });
+
+  test('Opens and closes drawer in uncontrolled mode', () => {
+    const { wrapper } = renderComponent(<AppLayout drawers={[testDrawer]} />);
+    expect(wrapper.findActiveDrawer()).toBeNull();
+
+    wrapper.findDrawerTriggerById('security')!.find('span')!.click();
+    expect(wrapper.findActiveDrawer()).not.toBeNull();
+
+    wrapper.findActiveDrawerCloseButton()!.click();
+    expect(wrapper.findActiveDrawer()).toBeNull();
   });
 
   testIf(size !== 'mobile' && theme !== 'classic')('shows trigger button as selected when drawer opened', () => {
